Update jill.html
This program is to help students struggling with Algebra 1. It covers the "FOIL" method. Students can locate the 2 values in each problem, enter them into the text-fields, then when they hit submit, the answer to the problem will show up in the answer box.
**I am still in the process of getting the buttons to function correctly, but as far as formatting colors and text sizes, I am done.
I still need to also figure out how to increase the size if the text-fields and buttons to match the font-size/table-size.